requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# HTTP Client for callbacks
httpx>=0.26.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dateutil>=2.8.0
